refactor:feature/sidebar 하위 폴더 생성
